change deploy script

+++ deploy.sh
@@ -1,4 +1,7 @@
-git pull
+if git pull | grep -q 'Already up to date.'; then
+  exit 1;
+fi
+
 composer dump-autoload --no-dev --classmap-authoritative
 rm var/cache -R
 php bin/console doctrine:schema:update --force
